# documentation/static/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Provider-neutral HTTP and server-sent-events API for running agent turns, reading retained messages, resolving tool decisions, and managing subagents.",
        "title": "AgentCLI HTTP API",
        "contact": {
            "name": "AgentCLI API support"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/confirmations/{confirmationID}/decisions": {
            "post": {
                "description": "Answers a tool-authored informational Yes/No confirmation. This is independent of permission policy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmations"
                ],
                "summary": "Resolve a main-agent tool confirmation",
                "operationId": "resolveMainAgentConfirmation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Confirmation request ID",
                        "name": "confirmationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Confirmation answer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agentcli.ConfirmationDecisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.ConfirmationDecisionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/permission-mode": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Read the current permission mode",
                "operationId": "getPermissionMode",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PermissionModeResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Changes live tool permission evaluation and emits a permission-mode event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Change the permission mode",
                "operationId": "setPermissionMode",
                "parameters": [
                    {
                        "description": "New permission mode",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agentcli.SetPermissionModeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PermissionModeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/permissions/{permissionID}/decisions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Resolve a main-agent tool permission",
                "operationId": "resolveMainAgentPermission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission request ID",
                        "name": "permissionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission decision",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agentcli.PermissionDecisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PermissionDecisionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/sessions/{sessionID}/events": {
            "get": {
                "description": "Replays and follows every main-agent turn in a session, including queued lifecycle records, response-scope boundaries, and turns created automatically from subagent results. The session cursor is independent from each runtime event sequence.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Event streams"
                ],
                "summary": "Stream retained and live session activity",
                "operationId": "streamSessionEvents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resume after this session cursor",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Resume after this session cursor when after is omitted",
                        "name": "Last-Event-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One SSE data payload; the HTTP response remains open for more records",
                        "schema": {
                            "$ref": "#/definitions/agentcli.SessionEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{sessionID}/messages": {
            "get": {
                "description": "Returns the provider-neutral transcript, including user, assistant, tool-call, and tool-result messages.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "List retained session messages",
                "operationId": "listSessionMessages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.MessagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{sessionID}/turns": {
            "post": {
                "description": "Starts immediately when the session is idle or queues FIFO when another turn is active. Send Accept: text/event-stream to stream an admitted turn instead of receiving JSON.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "Turns"
                ],
                "summary": "Start or queue a main-agent turn",
                "operationId": "startMainAgentTurn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Turn input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agentcli.StartTurnRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/agentcli.StartTurnResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/agentcli.StartTurnResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{sessionID}/turns/{turnID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turns"
                ],
                "summary": "Read main-agent turn status and result",
                "operationId": "getMainAgentTurn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Turn ID",
                        "name": "turnID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.TurnResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "agentcli.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/agentcli.APIError"
                }
            }
        },
        "agentcli.ConfirmationDecisionRequest": {
            "type": "object",
            "required": [
                "answer",
                "call_id",
                "session_id",
                "turn_id"
            ],
            "properties": {
                "answer": {
                    "$ref": "#/definitions/confirmation.Answer"
                },
                "call_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
        "agentcli.ConfirmationDecisionResponse": {
            "type": "object",
            "properties": {
                "decision": {
                    "$ref": "#/definitions/agentcli.ConfirmationDecisionResponseValue"
                }
            }
        },
...
        "agentcli.MessagesResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.MessageResponse"
                    }
                }
            }
        },
...
        "agentcli.PermissionDecisionRequest": {
            "type": "object",
            "required": [
                "call_id",
                "decision",
                "session_id",
                "turn_id"
            ],
            "properties": {
                "call_id": {
                    "type": "string"
                },
                "decision": {
                    "$ref": "#/definitions/permission.DecisionType"
                },
                "session_id": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
        "agentcli.PermissionDecisionResponse": {
            "type": "object",
            "properties": {
                "decision": {
                    "$ref": "#/definitions/agentcli.DecisionResponse"
                }
            }
        },
...
        "agentcli.PermissionModeResponse": {
            "type": "object",
            "properties": {
                "mode": {
                    "$ref": "#/definitions/permission.Mode"
                },
                "previous": {
                    "$ref": "#/definitions/permission.Mode"
                }
            }
        },
...
        "agentcli.SetPermissionModeRequest": {
            "type": "object",
            "required": [
                "mode"
            ],
            "properties": {
                "mode": {
                    "$ref": "#/definitions/permission.Mode"
                }
            }
        },
        "agentcli.StartTurnRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "description": "User message for the new turn.",
                    "type": "string"
                },
                "turn_id": {
                    "description": "Optional caller-defined idempotency identity.",
                    "type": "string"
                }
            }
        },
...
    }
}